Move to using a common HazeNode impl
