+++ haze/src/commonMain/kotlin/dev/chrisbanes/haze/Platform.kt
@@ -6,9 +6,4 @@ package dev.chrisbanes.haze
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.node.CompositionLocalConsumerModifierNode
 
-internal expect fun createHazeNode(
-  state: HazeState,
-  style: HazeStyle,
-): HazeNode
-
 internal expect fun CompositionLocalConsumerModifierNode.calculateWindowOffset(): Offset
