+++ haze/src/skikoMain/kotlin/dev/chrisbanes/haze/HazeNode.skiko.kt
@@ -1,0 +1,7 @@
+package dev.chrisbanes.haze
+
+import androidx.compose.ui.node.CompositionLocalConsumerModifierNode
+
+internal actual fun CompositionLocalConsumerModifierNode.updateRenderEffect(effect: Effect) {
+
+}
